feat: Add a log shipper for Fly.io deployments.
This log shipper connects to our Tailscale network ("Tailnet") so that
we can securely ship logs to our logging instance, because a) Loki
doesn't support any form of authentication and b) Cloudflare Access's
token-based auth method for apps isn't supported in Vector, which
we'll be using to ship the logs.

Note that this commit doesn't enable Vector yet. I'm just vetting this
Tailscale mechanism first.