refactor: Fix/workaround HMR issue.
What we initially thought was a Storybook issue (see
https://github.com/hackworthltd/primer-app/issues/548) turns out to be
some sort of Vite HMR problem with the way we were exporting our
component classes. See
https://github.com/hackworthltd/primer-app/pull/546#issuecomment-1272707729
for details of how the issue manifested in Storybook, and
https://github.com/hackworthltd/primer-app/issues/548#issuecomment-1282252834
for details of how one could trigger it from `pnpm watch` on the
application itself.

Somehow this issue only manifested once we merged our components into
a single project alongside the application code in
https://github.com/hackworthltd/primer-app/pull/547, possibly because
we're now importing the components directly from the filesystem,
rather than via an intra-workspace package dependency.

It's not precisely clear what is happening here, nor why the changes
in this commit fix it, but given the constant parade of module- and
import-related issues endemic to the JavaScript ecosystem, frankly I'm
not sure we really care to dive in and figure it out.

Suffice to say that, in addition to fixing the issue at hand, this
commit also simplifies the definition of our components by removing a
boilerplate-y `index.tsx` file that simply re-exports the definitions
defined in the "home" `.tsx` file.

Closes https://github.com/hackworthltd/primer-app/issues/548