fix: More Tailscale-for-Fly.io-logging fixes.
Specifically:

* Make a note that the Tailscale auth keys we create for this purpose
must have the "Reusable" flag set. See
https://github.com/tailscale/tailscale/issues/5982

* Make sure the entrypoint script can find `sleep` (!)

* Run `tailscale logout` on exit so that the Tailscale hostname will
be reusable. Note that, due to the way Tailscale rolls out
deployments (by default, anyway), I expect that the new instance will
start before the old one is killed, and therefore the canonical
hostname might not be available anyway when the new instance is
started. This may lead to a "ping-pong"-ing of hostnames like
`hackworth-codes-logging` and `hackworth-codes-logging-1`. It's
possible that we may be able to change the Fly.io rollout mechanism to
always kill the old instance before starting the new one, if this
becomes a problem and we're willing to have a brief logging
interruption during deployments.