fix: Don't use `primer-app` gitrev to tag the logging Docker image.
The Nix code to generate the logging Docker image was copied from the
same for the `primer-service` Docker image in the `primer` repo. For
`primer-service`, we want to tag the Docker image with the git rev
that was used to generate it, to make it easier to debug deployed
`primer-service` instances and to know exactly what git rev a deployed
server corresponds to.

For the logging service, we don't need to be so exacting, and we can
just use the default Docker image tag mechanism, which I believe is
based on the image's SHA. In fact, as we would expect the dependencies
of the logging service image to change relatively
infrequently (certainly no more often than the `primer-app` `nixpkgs`
pin, in the worst case), we'd prefer *not* to redploy this logging
image unless it has actually changed; whereas the previous method
would cause it to be redeployed on every merge to `main`, regardless
of whether anything had actually changed or not.

(Also update the local deployment script to synchronize it with these
changes. Note that this script is only meant to be used when we need
to deploy manually, as production deployments are handled by a GitHub
workflow.)