feat: Use actual last-modified times from the service.
Most of the hard work here was done server-side, as we already had
support for last-modified times in the frontend UI, using placeholder
values in lieu of the actual time from the backend.

However, one front-end surprise was that Orval doesn't automatically
convert strings to `Date`s for us: we have to do this ourselves via
some fairly ugly Axios "interceptor" code. See
https://orval.dev/reference/configuration/output#usedates