feat: Define a code typeface and use it where code appears in the UI.
Per our UI design spec, I've chosen a serif font for code, but I've
used Adobe Source Serif rather than what's in the spec. We can
bikeshed this choice, and other aspects of the font family, later.
However, we would like to be able to distribute the fonts we use,
rather than going through a font CDN, as font CDNs are notorious for
tracking users. This probably means we'll need to choose a font under
an open source license (and ditto for any other fonts we use).