fix: Un-escape font family names.
The Tailwind docs explicitly tell you that Tailwind doesn't escape
font names, and therefore you need to double-quote or otherwise escape
font names with, e.g., spaces in the name. Ref:

https://tailwindcss.com/docs/font-family#customizing-your-theme

However, I noticed that when we do that, Safari does *not* render the
font face correctly! It ignores the font family setting and chooses
the next font in the list, despite the fact that the Safari developer
tools show the proper font being selected. (There are no console
errors, either.)

For the record, Firefox is not affected by this issue, and renders the
correct font when the name is escaped.

When I remove the escaping (double-quotes, in our case), everything
works properly on Firefox, Safari, and Chrome. I'm not sure why it's
not necessary despite the Tailwind warning. Perhaps Vite is doing this
for us during the bundling process?

In any case, I've removed the double-quotes in this commit, as that
works in all cases according to my testing.