feat: Use Inter as the UI font (sans).
This isn't a final decision, but it meets our criteria here:
https://github.com/hackworthltd/primer-app/issues/580. (Also, note
that Inter *is* the font that Ann chose for UI elements in her design
spec.)

Note that on macOS and iOS, this font is barely distinguishable from
the system default (which I believe is SF Pro). It's a bit heavier,
but I had to zoom in to see it. If we stick with this font, it might
be better to use the system default on macOS and iOS as I'm sure
Apple's font has better language and glyph coverage.

Also note: this is version 3.19 of Inter. A new 4.0 release is coming
soon (currently in beta), but it's not quite ready yet.