chore: Switch branding from "Primer" to "h:".
This reflects the company's new branding, and the decision to call the
product "Hackworth Codes" rather than "Primer."

Also:

* Use Source Serif Semibold as the branding font. Note that when I
tried to use the Source Serif 4 variable font for this (with font
weight 600), it did not render correctly, so here I've had to resort
to using the fixed semibold variant. I'm not sure why the variable
font doesn't work. It's possible that I don't know enough about
variable fonts and I'm doing something wrong in the `@font-face`
settings, but my understanding is that it should "just work."

* Remove the `size` property for the branding. We're treating this
more like an icon than text: it just happens to be rendered as text.
Therefore it probably does not need to be responsive. We can add back
a breakpoint or two later, if necessary.

* Replace the bogus `<text>` HTML element in the branding component to
`<div>`.