chore: Sync up Tailwind config with Ann's final color document.
A few notes:

* Our Tailwind config inverts the blue primary and secondary colors
wrt. the document. (By my recollection, I think this was intentional
and had something to do with the corresponding hover colors.)

* Ann's document lists "red secondary" twice, once as #cd3764 and once
as #f1685e. I believe #cd3764 was added late in the process and
therefore I've called it "red tertiary" in this commit.

* Ditto for "grey secondary," but in this case, the hex RGB value
given for the 2nd occurrence of "grey secondary" does not match the
color swatch. The hex value is the same as the 1st occurrence of "grey
secondary" and therefore I've used the hex RGB value I measured by
sampling the swatch in the document (which is #d4d4d4). I've called
this "grey quaternary" in the Tailwind config.