fix: Use `ring` rather than `outline` for node... outlines.
I had been wondering separately what was the difference between `ring`
and `outline` and finally stumbled across an explanation of why `ring`
was added in the first place:

https://github.com/tailwindlabs/tailwindcss/discussions/7649

It turns out that on Safari, outlines don't follow the border of the
element they're outlining: they are always squared off. Tailwind CSS
addressed this by adding `ring`, so for our purposes, this is what we
should be using.