feat: Style labels per UI design spec (mostly).
There's still some work to be done here, but this is a good
approximation of the design spec.

Among the things that remain to be done are:

* We don't draw any labels on most syntax nodes.

* The size of the label and the contrast need some work.

* The single-glyph labels render like relatively flat ellipses rather
than nice circles like the 3-glyph labels do. We might be able to fake
this by padding out the single-glyph labels to 3 (spaces on either
side), or perhaps with a bit of min-width magic.