fix: Use a solid fill color for "syntax" nodes.
With this commit, we pretty closely match the UI design spec for tree
styling, with a few exceptions:

* Holes are not rendered as a circle.

* We don't have an `=` node for `let`s.

* Value constructor and function `apply` nodes read just `apply`
rather than `apply function`. We need to think about this a bit in the
context of patterns, where `apply` only ever appears with a value
constructor and we don't want students to get confused.

Note that there are a few node types that weren't specified in the UI
design spec (e.g., type constructors). At the moment, these are
rendered in black, and we'll need to choose some colors for these.

Also, nodes that appear in patterns are also rendered in black. I'm
guessing this is a leftover from a debate we had a few months ago
about whether patterns should be treated specially other than the box
drawn around them. This will be corrected in a subsequent commit.

Finally, the body content I've chosen here for some syntax
nodes (e.g., `function type`) should be considered just placeholders
for now. We should finalize the terms we use for these at a later
date. (One complication is that they should be reasonably short, so
that they don't take up too much space in the tree.)