feat: show imported types and definitions in sidebar
We currently just show the baseName of each. This could result in name
clashes if one has multiple imported modules defining the same name.
I.e. both 'M1.foo' and 'M2.foo' will be displayed as 'foo'. Since there
is currently no way to import a module (i.e. we are stuck with the two
default 'Primitives' and 'Builtins'), this is not practically a problem.