refactor: Rename `ActionButtonList` to `ActionPanel`
We also remove some comments about `ActionPanel` in `src/components/ActionButton/ActionButton.stories.tsx`. We've had the button list for a while, including its own stories, so there's nothing remaining to do. We now elide a background in the `ActionButton` stories so that we don't have to worry about it falling out of sync and being misleading as to the appearance of `ActionPanel`.