refactor: Split up `convertTree`
This is now more principled and should be easier to understand and modify. We have an initial step which augments nodes with all the data needed for ReactFlow to render them, then a second step (now much more generic) for converting trees to graphs.

We keep `augmentTree` as the first step, since it gives us information which we'll later need for layout, such as node dimensions. While we could easily structure things differently now, due to most node sizes being static, this would be difficult in future when we want to make them adapt to text length.

This refactor will also make it much easier to adapt to using any layout algorithms which require a tree as input, rather than a graph.