feat: Use Tidy algorithm for tree layout
We switch from Dagre for a number of reasons, including aesthetics and library maintenance status. But most importantly, by being tree-based, rather than working on arbitrary graphs, Tidy trivially gives us control over the left-to-right placement of sibling nodes, a feature _critically_ lacking in Dagre for our purposes. (An issue has been open for eight years, with the functionality frequently asked about (there are even multiple PRs purporting to fix it, though none with any real documentation): https://github.com/dagrejs/dagre/issues/112. We might have realised sooner that this would be an issue, as our translation from tree to graph clearly does nothing to preserve child order, if not for the fact that somehow the order, despite being effectively non-deterministic (our experiments have shown it's not even lexicographic on IDs or anything else one might expect), is somehow correct a significant majority of the time. It's only now that we are able to build programs in-app again that this limitation has become clear.)

Note that we do not _yet_ remove Dagre fully. It's still used here for layout of nested graphs (i.e. pattern boxes), pending a refactor.

We now layout definitions "one tree at a time", by necessity, as opposed to the "combine everything in to one graph first" approach we used with Dagre. This results in, for now, larger gaps between trees. But this is something we wanted to explore anyway, regardless of Tidy, since the old trees often blurred together a little too much because of the lack of spacing. We can improve this in future by using a more sophisticated packing algorithm, whereas currently we just draw a bounding rectangle around each tree then juxtapose horizontally.

Note also that how we build & import `@zxch3n/tidy` is a bit of a hack:

- We build the package using our https://github.com/hackworthltd/tidy fork, via `nix-build -A packages.x86_64-linux.tidy`.
- We create a `dist` branch of that repo and check into it the `dist`, `wasm_dist`, and `package.json` files from the `nix-build` step.
- We add the package to `primer-app`'s `package.json` via a special GitHub URL; e.g., `github:hackworthltd/tidy#dist-rev`

This works, but it's obviously not maintainable long-term. If we stick with Tidy, we'll need a different approach. (Note that using GitHub Packages' npm repository is a non-starter for us: see https://github.com/hackworthltd/primer-app/pull/638#issuecomment-1329120132)