feat!: actions for inserting primitives
This just bumps the primer pin (to my primitives branch) since this
logic is implemented in the backend. There are also some small changes
to adapt this project to the changed API.

BREAKING CHANGE: this enlarges changes the set of available actions,
which are stored in the log in the DB. Since it is only an increase, old
databases will still be acceptable. However, DBs created with this PR
will not be able to be loaded into an older version. Also, note that the
new version of primer changes the api, so a new frontend cannot talk to
an old backend and vice versa; however, this has no impact on the DB.