feat: show evaluation results in sidebar (stage 2)
We now show it as a tree, with dummy definition and type nodes. In the
sequel we will avoid the need for this dummy.

Note that we do not make it clear whether the evaluation completed or
timed out (i.e. whether we have hit a normal form or not).