chore: migrate/bump react-flow-renderer -> reactflow
The 'react-flow-renderer' package is deprecated in favor of 'reactflow'.

This also bumps the version from 10.3.17 to 11.3.2.

Note that there are still references to react-flow-renderer in
pnpm-lock.yml. I will fix these in a follow up commit by regenerating
the lockfile and bumping bounds of all dependencies.

This also does some minor changes to fix the build. The output is styled
somewhat differently - this will be addressed in subsequent commits.