feat: New session list previews.
These are styled now as cards, with buttons for favoriting &
sharing (not yet hooked up, as the backend lacks support for this).

I have some concerns about how busy the UI is now, but I think this is
better than the alternatives I could come up with, which are:

1. Hide favoriting and sharing behind a long press, which it appears
Apple do not want to support on mobile Safari:
https://github.com/hackworthltd/primer-app/issues/188#issuecomment-1336660240

2. Single-icon favorite and share buttons, but these are not
particularly informative since they don't have any explanatory text.

I'm open to option 2, as they would take up less screen space and
there would be fewer words on the page, but it's not as discoverable
as the solution implemented here, which includes the words "Favorite"
and "Share" along with the iconic representations. We could support
both, and only display the icons in a more advanced mode, once the
student has some experience using the application.

Note that this commit introduces a new style of button, a sort of
built-in style for cards. I played around with our standard `UIButton`
instead of this new style, as I'm reluctant to add new UI design
language in general; but I think in this case, because we're adding a
major new UI element in the cards themselves, that it makes sense that
they have their own button style. The `UIButton` looked awful in this
context, IMO.