fix: Fix Safari scrolling issue caused by new session list impl.
See the GitHub PR for details of the problem I was trying to solve
with the original, slightly wonky implementation, and an example of
how it broke scrolling on Safari:

https://github.com/hackworthltd/primer-app/pull/660#issuecomment-1339580196

The implementation in this commit resolves the Safari scrolling issue,
but does suffer from the "cards aren't the same size" problem when
dates wrap. However, as mentioned in the thread cited above, given
that we now render dates with `text-xs`, that particular issue is very
unlikely to occur, whereas the Safari scrolling issue is omnipresent.