fix: Fix major performance regression in session list on Safari.
The previous commit introduced a performance regression on Safari,
where the buttons on each card in the session list are *extremely*
laggy when highlighting on hover, sometimes taking up to 2 to 3
seconds to respond, and apparently queuing up a series of delayed UI
events that take quite awhile to catch up. This bug is present on both
Safari for macOS and iOS.

What's particularly frustrating is that the previous commit was
necessary to fix a Safari scrolling issue.

Anyway, to fix the performance issue, we make the top-level div in the
sessions page `relative`, and remove all nested `relative` positions
in the child elements. This brings Safari back to Firefox and Chrome
performance, and preserves the fix for the scrolling issue.

Where did all these `relative` classes come from?

* Most of the Tailwind classes for the cards came from this Tailwind
UI component, which uses `relative` for each button on the cards:
https://github.com/hackworthltd/tailwind-ui/blob/3c82675b8928c622d821119265d6d83d982dac82/react/components/application-ui/lists/grid-lists/contact_cards.jsx#L44

* The `relative` on each `<li>` pre-dates these session list changes.

As far as I can tell, removing all of these nested `relative` classes
makes no difference to how this particular UI is rendered.

Note that neither Firefox nor Chrome need this workaround: their
performance was fine with the previous commit.