fix: Use selection from backend
This avoids errors after performing an action, such as trying to access a node ID which no longer exists.

Note that we still allow the selection to be set in the frontend independently, by clicking on a node. But we now override this selection when a new `Prog` comes in from the backend. Previously, the `Prog`'s selection was ignored.