feat: Delete sessions via the delete button on session cards.
Note that this commit breaks the Storybook stories for the
`SessionPreview`, `SessionList`, and `SessionsPage` components,
because when we run in Storybook, we can't (and don't want to) rely on
a running Primer service instance, which is what `useQueryClient`
expects.

We'll need some mocks to make this work again. Adding support for that
is a fairly major change (we don't currently have any mocks), and
therefore I'd like to do that as a separate commit, which will follow
this one.