fix: Use deep equality for comparing global names
In JavaScript, `==` on objects is reference equality. This caused a bug where no selection would be shown upon creation of a new session or definition. This didn't manifest until recently because the selection was always created in the frontend and thus the objects _did_ have the same identity when we expected them to.

Fortunately, this was the only place in the codebase where we tested equality of objects. So we are not exposed to other bugs with the same cause.