fix: Add React Query mocks for session preview cards.
This fixes the Storybook for session previews by mocking the Primer
service endpoint for session deletion. Note that the mock doesn't
bother to check that the session ID is legitimate. (This would be
possible with some more work, and may be worth it if we want to use
Storybook to test error handling.)

Note that we need to set up this mock for every Storybook story that
includes the `SessionPreview` component. It may be possible to do this
at the Storybook project level, if we're happy to mock some responses
that could be used for all stories.