feat: Delete sessions via the delete button on session cards.
Here we have explicitly chosen to implement the per-session deletion
callback as an on-click handler that we thread all the way down the
session list component hierarchy to `SessionPreview`, where it will be
invoked (as that's where the specific session ID is known). The
advantage of this approach is that we can keep all the React Query
code in the top-level `ChooseSession` component.

(See https://github.com/hackworthltd/primer-app/pull/672 for an
alternate approach.)