fix: Refetch when the last session on the last page is deleted.
Previously, the frontend would just display an empty page (and a
nonsensical "Showing m of n sessions" footer, with m > n) when this
condition occurs.

This fix also handles the case where the frontend somehow requests a
page beyond the last page of results. This could happen if, for
example, the student tries to advance to the last page of results, but
another client has, in the meantime, deleted the sessions that would
have otherwise been on the last page (i.e., a race between 2 clients).

We could also handle this at the API level, but it's quite
straightforward to do here with no backend changes required: it's just
2 simple sanity checks on the API result.