fix: Globally disable React Query caching mechanism
By default, `useQuery` hooks return cached data until the server responds. This was causing bugs due to us unintentionally making use of this stale data.

For example, the `App` component was being rendered more than once because of `useGetProgram` first returning a stale program, and later a fresh one from the server. But we were implicitly assuming that it would only return once, and using its value to initialise a state hook in `AppProg`. This meant that when a session was loaded, the program state would often not reflect recent edits.

Given that we hadn't until recently realised that React Query was doing this, it's very possible that this behaviour was causing further undiscovered bugs. So, for now at least, we take the safe, if potentially non-optimal, option of disabling it completely.

For a full explanation of React Query's behaviour, see https://tanstack.com/query/v4/docs/guides/caching.