fix: reactflow background state sharing
This fixes a bug-in-waiting: if one has multiple flows on one page, they
each need to be given an unique id (in reactflow 11), otherwise one of
them will control the state of the background for all the flows.

This is not currently a problem, but will be shortly when we display the
result of evaluation as a tree.