fix: clicking '+' in T&D Types list does not add a Definition
We split the click handlers for the TypesAndDefinitions and Sidebar
components into two sets: one for the types and one for the terms. We
also take the opportunity to refine the types: click handlers for each
individual (type/term) definition take arguments indicating which one
was clicked, but the handlers for the '+' (add a definition) button need
no arguments.