chore(nodejs): Bump dependencies.
Upgrade to Vite 4.0 again, as the issue I'm having with local
Storybooks appears to be unrelated to the 4.0 upgrade. (See
https://github.com/vitejs/vite/issues/9986)

While we're at it, remove the Vite cachedir override, in case this is
related to https://github.com/vitejs/vite/issues/9986. We originally
did this in order to accommodate a non-writable node_modules in the
future, but that seems very unlikely to happen anytime soon.