feat: Allow setting arbitrary additional class names on `UIButton`.
We need a bit more flexibility from our `UIButton`. Adding a
`className` parameter that can be used as it is for other React
components provides maximum flexibility.

(Perhaps this is too flexible? If this ends up being a footgun, we can
revisit this decision, and just add one-off parameters for the
specific flexibility we need.)