fix: Fix the TypeScript build.
82d46100b0bf72df6df5c3d7ecc90b3404862dd0 broke the TypeScript build
because `UIButton`'s new `className` property has type `string |
undefined`, but Storybook's controls are too limited to express that
type, and can only be `string`. Because we're enabling
`exactOptionalPropertyTypes` in our TypeScript config, TypeScript
fails when it sees the mismatch.

However, TypeScript shouldn't even be looking at our Storybook stories
when building the app! This was an omission: we should be excluding
stories from the app build, and we just haven't noticed until now.
This is easy enough to do: just add `*.stories.tsx` to TypeScript's
`excludes` when building the app, which we can do by extending a base
TypeScript config that's shared by both Storybook and the app.

[Note that I'd very much prefer to name the TypeScript app config
`tsconfig.json` and the StoryBook-specific one
`tsconfig.storybook.json`. However, I could not manage to make that
work. The `vite-tsconfig-paths` plugin doesn't work properly when I
change its `root` property to point to a Storybook-specific TypeScript
config. It's possible I'm doing something wrong, but I've been
struggling to make this work for several hours now, and have given up.]