feat: Simple dialog to prompt for new definition name.
The implementation currently has the following limitations:

* No error handling: if the backend returns an error on definition
creation, there's no indication shown to the student.

* The dialog icon is the same as that for the new session name dialog.
We should choose a different one, but Heroicons doesn't have an
obvious choice, so I've reused the new session one for now.

* Limited validation: we check for an empty name, but nothing else.
If, for example, the name provided is the same as another definition
in the same module, we rely on the backend to catch that. We could
check this client-side to prevent a round trip, but once we have
proper error-handling, it's not clear whether it will be worth it.

* DRY: the implementation is a copy-paste of the new session name
dialog component. We could DRY quite a bit of this, but I want to run
the styling past the rest of the team first, and they're not currently
available. We can come back to this later.