feat: Use Zod to validate session names in the modal dialog.
In addition to the previous empty-name check, we now additionally do
the following:

* Trim whitespace from the string that's sent to the server.

* Add a client-side check to enforce the server-side maximum session
name length of 64 (whitespace-trimmed) characters.

We may want to remove the 64-character check in the longer term, but
I've included it here mainly as a way to demonstrate the sorts of
checks we easily do with Zod.

Note that Zod's schema validation is a bit odd, or at least it is if
you use `.trim()` on strings: it is imperative, not delcarative,
meaning that 1) it will actually modify the data returned by the
form (i.e., the string in the form data will have been trimmed), and
it's order-dependent, in that the min/max length checks will not
consider the trimmed string *unless* the `.trim()` check comes before
the chained min/max checks. I've left a comment in the code about this
as a warning to the reader, but it won't be feasible to leave this
comment everywhere we use Zod.