doc: Rewrite error messages in typedef form.
It's a bit tricky coming up with form errors that are descriptive, but
not too long. I've now errored (hah) on the side of stating the
problem succinctly, but not elaborating on the rule that's being
violated. In subsequent work, we can add an info icon to each error
message that will provide a more comprehensive explanation of the
problem, including specific rules, links to the language spec,
examples, etc.