fix: Add `type="button"` to typedef form add/remove buttons.
Per
https://github.com/react-hook-form/react-hook-form/discussions/9766#discussioncomment-4694254,
this fixes the issue where, in the typedef form, clicking the "add"
button prior to submitting the form would trigger a validation and
then cause validations to run on `onChange`.

Now the form acts just like a simple, static `useForm` form where
validation doesn't begin until the submit button is pressed, and is
thereafter `onChange`.