feat: Add a button type property to `UIButton`.
As evidenced in the previous commit, all buttons are not created
equal, and so we will need to specify a `UIButton`'s `type` attribute.

The default value for this new property is `"button"`, as that's the
value that was previously hard-wired and therefore doesn't require any
changes in the existing code base.

Notes:

* Default type `"button"` is different than in HTML, where it's
`"submit"`. However, that seems like a legacy decision that predates
using HTML `<button>`s in web app UIs.

* We do not support the HTML `"reset"` attribute value, which,
according to MDN, "tends to annoy users." (Ref:
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-type)