ci: Fix Cloudflare Pages deployment workflow.
In b7eceecddaa0faa6b60f583814eb155aebfdcb25, I changed the app build
workflow to run only on pushes to main, and pull requests, as a way to
reduce the number of builds we perform. (It was previously run on any
push.) However, I'd forgotten that the cloudflare/pages-action GitHub
Action only works well with `push`.

Therefore, this commit switches back to running builds on every push.