ci: Switch Chromatic deployment back to `push` trigger.
For whatever reason, Mergify can't associate `workflow_run`-triggered
GitHub workflows with the merge queue PRs it creates, so without this
reversion, it thinks the Chromatic deployment check hasn't run, even
when it has.

It's not really a big deal to change back to `push` for this
deployment. The main impetus for using `workflow_run` for our
deployment workflows was to ensure that we didn't push to Fly.io until
we were certain that the frontend build workflow had succeeded.
Deploying to Chromatic on every push (even ones that fail the build)
is a bit wasteful, but it's not the end of the world.