ci: Use GitHub deployments concurrency groups for Fly.io deployments.
This ensures we don't try to deploy more than one commit to `main` at
a time. Note that we *do* permit concurrency within the different apps
we deploy to Fly.io. In this commit, there are 2 groups: one for the
logging container, and one for the app container (i.e.,
`primer-service`).