ci: Deploy each component to a separately-named environment.
For each merge to `main`, we perform 3 different deployments, each to
a different component of the production service:

* The frontend code (to Cloudflare Pages);

* The backend service (to a Fly.io app named hackworth-code).

* The backend logging service (to a Fly.io app named
hackworth-codes-logging).

Prior to this commit, our GitHub CI workflows created the following
environments (and deployments):

* "Production", via the cloudflare/pages-action GitHub Action.

* "production", for both Fly.io deployments.

This commit creates a different environment for each Fly.io
deployment, one named "Production (logging)", and the other named
"Production (service)". This will make it possible to track each
deployment separately in the GitHub UI.

(Note that we have no control over the environment name used by the
cloudflare/pages-action, as far as I can tell.)

This commit also changes the URLs for the Fly.io deployments to their
Fly.io app pages, rather than to the services' URLs/endpoints, as the
latter are not very useful on their own.