feat: Layout right-children to the right
We achieve this by telling Tidy that they are children of the node which we want them to be positioned below. This is the node reached by traversing up the tree until the last edge followed is a _non-right_-edge (NB. implementation is simpler than this definition would suggest). In practice, given that right-children are currently always pattern boxes, this means the parent of the "match" node.

This commit includes a pretty significant rewrite of `layoutTree`. Our old implementation was not flexible enough to support dummy edges (i.e. supplying Tidy a different topology to the "real" one), since we indexed edges by the ID of their node in Tidy, and couldn't separate this from an edge's Primer source ID.

The layout is still certainly not perfect, but the topology is at least now correct. We can make further adjustments in due course, either via upstream changes to Tidy, or performing more post-processing on its output.