fix: Traverse all nodes before edges in constructing layout maps
The previous implementation happens to be safe right now, since dummy edges added for patterns will have a source node which features even higher up the tree than the "real" source, so the node will have already been added. But the intention was to be more general, as is clear from the comment saying "we've already added _all_ nodes".