refactor: Inline `primerNodeClasses` etc.
They're only used in one place, and this change just makes the code easier to follow.

We remove `"ring"` from selected nodes, since it's already covered by `"ring-4"`. Tailwind/ESLint was able to alert us to this redundancy, where it wasn't before, since the classes are now used directly in the `className` attribute.