refactor: Define `PrimerNode`
As described in source comments, this is similar to ReactFlow's `Node`, but more type-safe. We are able to statically assert that the `data` fields of our nodes match what we tell ReactFlow to expect.

`Equals` and `assertType` are inspired by, respectively, https://github.com/microsoft/TypeScript/issues/27024#issuecomment-420227672 and https://github.com/microsoft/TypeScript/issues/27024.

We carry on using standard ReactFlow edges and handles, for now. Though we will likely do something similar there in future, once we require further customization. In fact, we now define a very simple `PrimerEdge` type for a small degree of forwards-compatibility.

This also puts us in slightly better shape to ditch ReactFlow for an alternative, should we wish to, since we now implement better abstractions in some ways ourselves, using ReactFlow's types less extensively.