refactor: Simplify `augmentTree`
We now add the extra nested graphs to the node at which they are nested, then extract them later, instead of tangling them up with `augmentTree` in a rather ad-hoc manner. This pushes slightly more work on to consumers, for the benefit of making it easier to follow what happens to nested graphs. We will likely soon common up `TreeReactFlowOne` with `TreeReactFlow`, at which point there will only be two consumers anyway.