feat: Add a Kustomize deployment for use by our Argo CD server.
When Argo CD is properly configured, it will look for this directory
on each PR and, if certain conditions are met on the PR, deploy the
this application into our Kubernetes cluster.

Note that this Kustomize app is intended only for testing, and is not
a complete & proper installation of the `primer-service` service.