make box background more transparent
looks better, but slightly annoying that we see the RF dots - ideally we maybe want to modify the colour itself? or we could maybe set a plain white background, then put the transparent yellow one on top of that

maybe we should actually do this for all nodes, not just boxes? we already do something similar for def-name