ci: Split frontend deployment into 2 workflows.
cloudflare/pages-action, which previously did not work for us on
`pull_request` GitHub workflow triggers, now works with @v1.4.0.
However, this version has been reverted upstream as of @v1.4.1 (which
we were previously using), because, ironically, it broke others' pull
request-based workflows.

Upstream is now working on a version that will hopefully work for all
parties:

https://github.com/cloudflare/pages-action/pull/70

In any case, we now split our frontend deployment GitHub workflows
into 2: one for pull requests, and one for pushes to `main`.
Previously, we had to build & deploy the frontend on *any* push event,
which created too many jobs and occasionally blew up in weird ways for
pull requests.

Note that we could do some DRY here, but a) GitHub makes that
remarkably non-ergonomic, and b) I think it's quite likely that we
will drop the frontend deployment on pushes to `main` soon, anyway, as
that should really be managed now in the `hackworth-code-gitops` repo,
instead.