refactor: Use z-indices to reflect the nesting level of graphs
In practice, given the trees our backend outputs, this will always be 0 or 1.

This makes no visible difference yet, but will be important once we re-add solid pattern backgrounds.

Note that we have to use ReactFlow's own notion of `zIndex`. Manually adding it to `style` or `className` conflicts with ReactFlow's styling and has no visible effect.