feat: Add solid background to pattern boxes
We use opacity, along with a solid white background behind the actual colored one, as a way of modifying the default flavor color. Given that, in practice, we only have one box flavor (patterns - yellow), we could have created a new colour instead.

NB. We once had solid white backgrounds for patterns, but this was reverted in a4068c1 since we couldn't work out how to set the z-indices correctly.