ci: Push container image when "Deploy service" is applied to PRs.
We now push the pinned `primer-service` container image to ghcr.io
when the "Deploy service" label is applied to a PR. This means we can
use Primer pins that have not yet been merged to the Primer repo's
`main` in `primer-app` PRs, enabling a new kind of workflow where
Primer backend development can be done simultaneously with
`primer-app` development; i.e., you can have an open PR on the Primer
repo, and an open PR on the `primer-app` repo which depends on the
Primer pin in the Primer repo PR. (Prior to this commit, this workflow
was not possible unless you ran everything locally, because the Primer
repo doesn't push container images to ghcr.io until they've been
merged to `main`, and therefore in-development Primer pins could not
be deployed by Argo CD.)

Fixes https://github.com/hackworthltd/primer/issues/912.

This issue is better dealt with in this repo, rather than in the
Primer repo. By doing development container image pushes in this repo
-- and only on-demand, rather than for every push -- we will create
many fewer development packages in ghcr.io, as compared to pushing
every Primer PR push.

Notes:

* This changes the package used by the Argo CD application in this
repo to `ghcr.io/hackworthltd/primer-service-dev`. Note that this
is *different* than the "production" `primer-service` ghcr.io package,
which is `ghcr.io/hackworthltd/primer-service`. The latter package is
the one used by our Argo CD production Primer deployment.

* Do *not* add a version of this workflow that pushes to ghcr.io on
merges to `primer-app` `main`. We do not want to deploy production
backends from this repo. (Eventually, we will not even deploy
production frontends from this repo.)

* If you apply the "Deploy service" label to a PR, then change the
Primer pin and you want to deploy the new pin, you will need to remove
the label, then add it again. We could *probably* fix this so that
once the label is applied, every subsequent pin change would
automatically push the container image, but it might be tricky, and
I'm not sure it's worth the extra effort.